requirements.txt
lyricsgenius==3.0.1
nltk==3.8.1
pandas==1.5.3
python-dotenv==1.0.0
scikit_learn==1.2.2
spotipy==2.23.0
seaborn==0.12.2
transformers==4.30.1
torch==2.0.1